Back-end terminated

+++ src/main/java/com/example/progetto_psw/services/PurchasingService.java
@@ -51,25 +51,28 @@ public class PurchasingService {
 
         for(PipDetails pipDetails : pipDetailsList){
             Optional<ProductInPurchase> pipFinded = productInPurchaseRepository.findById(pipDetails.getId());
-            if(pipFinded.isEmpty()) throw new InconsistencyCartException("PRODUCT_"+pipDetails.getPid()+"_NOT_EXIST_TRY_LATER");
+            if(pipFinded.isEmpty() || pipFinded.get().getProduct().getId() != pipDetails.getPid()) throw new InconsistencyCartException("PRODUCT_IN_CART_"+pipDetails.getId()+"_NOT_EXIST_TRY_LATER");
             Optional<Product> p = productRepository.findById(pipDetails.getPid());
             if(p.isEmpty()) throw new IllegalArgumentException("PRODUCT_"+pipDetails.getPid()+"_NOT_EXIST");
             Product product = p.get();
             if(product.getQuantity() < pipDetails.getQta()) throw new QuantityProductUnavailableException(product.getId());
             if(product.getPrice() != pipDetails.getPrice()) throw new PriceChangedException(product.getId());
-            ProductInPurchase pip = new ProductInPurchase();
-            pip.setQuantity(pipDetails.getQta());
-            pip.setPrice(pipDetails.getPrice());
-            pip.setPurchase(purchase);
-            pip.setProduct(product);
-            pip.setCart(u.getCart());
-            productInPurchaseRepository.save(pip);
-            purchase.getProductsInPurchase().add(pip);
-            product.setQuantity(product.getQuantity() - pip.getQuantity());
+            if(pipDetails.getQta() > 0){ // Se la quantità del prodotto nel carrello è nulla allora viene ignorato nell'acquisto. Tuttavia lo faccio rimanere nel carrello così se torna disponibile può essere acquistato
+                ProductInPurchase pip = new ProductInPurchase();
+                pip.setQuantity(pipDetails.getQta());
+                pip.setPrice(pipDetails.getPrice());
+                pip.setPurchase(purchase);
+                pip.setProduct(product);
+                //pip.setCart(u.getCart()); dato che il responsabile della relazione è pip, se gli settiamo il carrello, al termine pip sara' anche all'interno del carrello
+                productInPurchaseRepository.save(pip);
+                purchase.getProductsInPurchase().add(pip);
+                product.setQuantity(product.getQuantity() - pip.getQuantity());
 
-            // Procedo con la rimozione del vecchio pip che era nel carrello
-            cart.getProductsInPurchase().remove(pipFinded.get());
-            productInPurchaseRepository.delete(pipFinded.get());
+                // Procedo con la rimozione del vecchio pip che era nel carrello
+                cart.getProductsInPurchase().remove(pipFinded.get());
+                productInPurchaseRepository.delete(pipFinded.get());
+            }
+
         }
 
         return purchase;
